+++ src/test/java/kitchenpos/unit/application/MenuServiceTest.java
@@ -3,10 +3,10 @@ package kitchenpos.unit.application;
 import kitchenpos.application.MenuService;
 import kitchenpos.domain.*;
 import kitchenpos.infra.ProfanityClient;
-import kitchenpos.stub.MenuGroupStubRepository;
-import kitchenpos.stub.MenuStubRepository;
-import kitchenpos.stub.ProductStubRepository;
-import kitchenpos.stub.ProfanityClientStub;
+import kitchenpos.testdouble.MenuGroupStubRepository;
+import kitchenpos.testdouble.MenuStubRepository;
+import kitchenpos.testdouble.ProductStubRepository;
+import kitchenpos.testdouble.ProfanityClientStub;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
