+++ src/test/java/kitchenpos/unit/application/OrderServiceTest.java
@@ -3,9 +3,9 @@ package kitchenpos.unit.application;
 import kitchenpos.application.OrderService;
 import kitchenpos.domain.*;
 import kitchenpos.infra.KitchenridersClient;
-import kitchenpos.stub.MenuStubRepository;
-import kitchenpos.stub.OrderStubRepository;
-import kitchenpos.stub.OrderTableStubRepository;
+import kitchenpos.testdouble.MenuStubRepository;
+import kitchenpos.testdouble.OrderStubRepository;
+import kitchenpos.testdouble.OrderTableStubRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
