+++ src/test/java/kitchenpos/unit/application/ProductServiceTest.java
@@ -3,9 +3,9 @@ package kitchenpos.unit.application;
 import kitchenpos.application.ProductService;
 import kitchenpos.domain.*;
 import kitchenpos.infra.ProfanityClient;
-import kitchenpos.stub.MenuStubRepository;
-import kitchenpos.stub.ProductStubRepository;
-import kitchenpos.stub.ProfanityClientStub;
+import kitchenpos.testdouble.MenuStubRepository;
+import kitchenpos.testdouble.ProductStubRepository;
+import kitchenpos.testdouble.ProfanityClientStub;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
